% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_prior.R
\name{ar1_covariance}
\alias{ar1_covariance}
\title{Calculate AR1 covariance matrix}
\usage{
ar1_covariance(n_time, rho, noise_sd, intercept_sd)
}
\arguments{
\item{n_time}{Number of time points}

\item{rho}{AR1 correlation term}

\item{noise_sd}{Standard deviation of noise process}

\item{intercept_sd}{Standard deviation of intercept}
}
\description{
Calculate AR1 covariance matrix
}
