% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_prior.R
\name{delta_regional_posterior}
\alias{delta_regional_posterior}
\title{Approximate delta posterior based on regional testing data}
\usage{
delta_regional_posterior(test_df, I_quant, control, imperfect)
}
\arguments{
\item{test_df}{A data.frame with the following columns
\itemize{
  \item{Nt: }{Total number of targeted tests in region}
  \item{nt: }{Number of randomised tests returning positive in region}
  \item{M: }{Population of region}
}}

\item{I_quant}{Posterior quantiles of prevalence calculated by
\code{\link{randomised_testing_prevalence}}}

\item{control}{A list of control parameters defined via
\code{\link{get_control_parameters}}}

\item{imperfect}{Logical, whether to account for imperfect testing}
}
\description{
Approximate delta posterior based on regional testing data
}
