% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{get_control_parameters}
\alias{get_control_parameters}
\title{Set control parameters}
\usage{
get_control_parameters(
  delta_AR_rho = 0.99,
  delta_AR_sd = 1,
  delta_inter_sd = 10,
  I_prior_mean = 0.005,
  I_prior_sd = 0.025,
  I_prior_upper = 0.04,
  alpha_testing = 0.001,
  beta_testing = 0.05,
  I_max = 1e+06,
  nbin = 200,
  n_quant_approx_bias = 25
)
}
\arguments{
\item{delta_AR_rho}{Correlation in AR1 component of the temporal Gaussian process on delta}

\item{delta_AR_sd}{Standard deviation in AR1 component of the temporal Gaussian process on delta}

\item{delta_inter_sd}{Prior standard deviation on delta intercept}

\item{I_prior_mean}{Mean of truncated Gaussian prior on prevalence}

\item{I_prior_sd}{Standard deviation of truncated Gaussian prior on prevalence}

\item{I_prior_upper}{Upper limit of truncated Gaussian prior on prevalence}

\item{alpha_testing}{False positive rate}

\item{beta_testing}{False negative rate}

\item{I_max}{Maximum value of prevalence considered in the model}

\item{nbin}{The number of bins into which prevalence I is divided}

\item{n_quant_approx_bias}{Number of quantiles to use to approximate delta EB posterior}
}
\description{
Set control parameters
}
