% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_quantiles}
\alias{get_quantiles}
\title{Extract quantiles from a (normalised) probability distribution}
\usage{
get_quantiles(x, values, n_quantiles)
}
\arguments{
\item{x}{Vector of (normalised) probabilities}

\item{values}{Vector of corresponding values}

\item{n_quantiles}{Number of quantiles}
}
\description{
Extract quantiles from a (normalised) probability distribution
}
