% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logit_mom}
\alias{logit_mom}
\title{Gaussian moment-matching on logit scale}
\usage{
logit_mom(x, values)
}
\arguments{
\item{x}{Vector of log-probabilities}

\item{values}{Vector of proportions}
}
\description{
Gaussian moment-matching on logit scale
}
