% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalise_logprob}
\alias{normalise_logprob}
\title{Normalise a vector of log probabilities}
\usage{
normalise_logprob(x)
}
\arguments{
\item{x}{A vector of log probabilities}
}
\description{
Normalise a vector of log probabilities
}
