% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_prevalence.R
\name{prior_prevalence}
\alias{prior_prevalence}
\title{Calculate prior prevalence}
\usage{
prior_prevalence(test_df, control)
}
\arguments{
\item{test_df}{A data.frame with the following columns
\itemize{
  \item{Nt: }{Total number of targeted tests}
  \item{nt: }{Number of targeted tests returning positive}
  \item{Nr: }{Total number of randomised tests}
  \item{nr: }{Number of randomised tests returning positive}
  \item{M: }{Population of local area}
  \item{delta_prior_mean: }{Prior mean of bias parameter}
  \item{delta_prior_sd: }{Prior standard deviation of bias parameter}
  \item{alpha: }{Alpha parameter of beta-binomial prior on probability of infectious given PCR positive}
  \item{beta: }{Beta parameter of beta-binomial prior on probability of infectious given PCR positive}
}}

\item{control}{A list of control parameters defined via
\code{\link{get_control_parameters}}}
}
\description{
Calculate prior prevalence
}
