% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_prevalence.R
\name{randomised_testing_prevalence}
\alias{randomised_testing_prevalence}
\title{Calculate log posterior of prevalence based on data from randomised testing}
\usage{
randomised_testing_prevalence(test_df, control, imperfect)
}
\arguments{
\item{test_df}{A data.frame with the following columns
\itemize{
  \item{Nr: }{Total number of randomised tests}
  \item{nr: }{Number of randomised tests returning positive}
  \item{M: }{Population of local area}
}}

\item{control}{A list of control parameters defined via
\code{\link{get_control_parameters}}}

\item{imperfect}{Logical, whether to account for imperfect testing}
}
\description{
Calculate log posterior of prevalence based on data from randomised testing
}
