% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_prior.R
\name{specify_delta_prior}
\alias{specify_delta_prior}
\title{Specify prior on delta using targeted and randomised regional testing data}
\usage{
specify_delta_prior(test_df, control, imperfect = TRUE)
}
\arguments{
\item{test_df}{A data.frame with the following columns
\itemize{
  \item{Nt: }{Total number of targeted tests in region}
  \item{nt: }{Number of targeted tests returning positive in region}
  \item{Nr: }{Total number of randomised tests in region}
  \item{nr: }{Number of randomised tests returning positive in region}
  \item{M: }{Population of region}
}}

\item{control}{A list of control parameters defined via
\code{\link{get_control_parameters}}}

\item{imperfect}{Logical, whether to account for imperfect testing}
}
\description{
Specify prior on delta using targeted and randomised regional testing data
}
