% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_prevalence.R
\name{targeted_testing_loglik}
\alias{targeted_testing_loglik}
\title{Calculate log likelihood for targeted testing data}
\usage{
targeted_testing_loglik(test_df, I, control, imperfect)
}
\arguments{
\item{test_df}{A data.frame with the following columns
\itemize{
  \item{Nt: }{Total number of targeted tests}
  \item{nt: }{Number of targeted tests returning positive}
  \item{M: }{Population of local area}
  \item{p1: }{Probability of taking a test given infectious}
  \item{p2: }{Probability of taking a test given not infectious}
}}

\item{I}{Prevalence}

\item{control}{A list of control parameters defined via
\code{\link{get_control_parameters}}}

\item{imperfect}{Logical, whether to account for imperfect testing}
}
\description{
Calculate log likelihood for targeted testing data
}
